fix route to satisfy app router build config
