+++ app/api/auth/[...nextauth]/route.ts
@@ -1,60 +1,6 @@
-import NextAuth, { NextAuthOptions } from "next-auth";
-import CredentialsProvider from "next-auth/providers/credentials";
-import { compare } from "bcrypt";
-import prisma from "@/app/prismaClient";
+import NextAuth from "next-auth";
+import authOptions from "@/utilities/auth";
 
-export const authOptions: NextAuthOptions = {
-  session: {
-    strategy: 'jwt'
-  },
-  pages: {
-    signIn: "/login"
-  },
-  providers: [
-    CredentialsProvider({
-      name: 'Credentials',
-      credentials: {
-        email: {},
-        password: {}
-      },
-      async authorize(credentials) {
-        if (!credentials?.email || !credentials?.password) {
-          return null;
-        }
-
-        const user = await prisma.users.findUnique({
-          where: {
-            email: credentials.email
-          }
-        });
-
-        if (user && await compare(credentials.password, user.password)) {
-          return {
-            id: user.id,
-            email: user.email
-          };
-        }
-        return null;
-      }
-    })
-  ],
-  callbacks: {
-    async jwt({ token, user }) {
-      if (user) {
-        token.id = Number(user.id);
-        token.email = user.email;
-      }
-      return token;
-    },
-    async session({ session, token }) {
-      if (session.user) {
-        session.user.id = token.id;
-        session.user.email = token.email;
-      }
-      return session;
-    }
-  }
-};
-
-// Default export for NextAuth API route
-export default NextAuth(authOptions);
+const handler = NextAuth(authOptions);
+// Export handlers for GET and POST methods
+export { handler as GET, handler as POST };
