+++ app/api/saveBill/route.ts
@@ -1,7 +1,7 @@
 import { NextResponse } from 'next/server';
 import prisma from '@/app/prismaClient';
 import { getServerSession } from 'next-auth';
-import { authOptions } from '../auth/[...nextauth]/route';
+import authOptions from '@/utilities/auth';
 
 export async function POST(request: Request) {
   const session = await getServerSession(authOptions);
